d/aws_regions: Remove terraform-plugin-sdk implementation.
